+++ src/global/misc/src/MResponseMatrixO15.cxx
@@ -891,11 +891,11 @@ float MResponseMatrixO15::GetMinimum() const
 ////////////////////////////////////////////////////////////////////////////////
 
 
-float MResponseMatrixO15::GetSum() const
+double MResponseMatrixO15::GetSum() const
 {
   // Return the sum of all bins:
 
-  float Sum = 0;
+  double Sum = 0;
   for (unsigned int i = 0; i < m_AxesO14.size(); ++i) {
     Sum += m_AxesO14[i].GetSum();
   }
