+++ src/global/misc/src/MResponseMatrixO17.cxx
@@ -937,11 +937,11 @@ float MResponseMatrixO17::GetMinimum() const
 ////////////////////////////////////////////////////////////////////////////////
 
 
-float MResponseMatrixO17::GetSum() const
+double MResponseMatrixO17::GetSum() const
 {
   // Return the sum of all bins:
 
-  float Sum = 0;
+  double Sum = 0;
   for (unsigned int i = 0; i < m_AxesO16.size(); ++i) {
     Sum += m_AxesO16[i].GetSum();
   }
