+++ src/global/misc/src/MResponseMatrixO2.cxx
@@ -728,11 +728,11 @@ float MResponseMatrixO2::GetMinimum() const
 ////////////////////////////////////////////////////////////////////////////////
 
 
-float MResponseMatrixO2::GetSum() const
+double MResponseMatrixO2::GetSum() const
 {
   // Return the sum of all bins:
 
-  float Sum = 0;
+  double Sum = 0;
   for (unsigned int i = 0; i < m_AxesO1.size(); ++i) {
     Sum += m_AxesO1[i].GetSum();
   }
