+++ src/global/misc/src/MResponseMatrixO4.cxx
@@ -750,11 +750,11 @@ float MResponseMatrixO4::GetMinimum() const
 ////////////////////////////////////////////////////////////////////////////////
 
 
-float MResponseMatrixO4::GetSum() const
+double MResponseMatrixO4::GetSum() const
 {
   // Return the sum of all bins:
 
-  float Sum = 0;
+  double Sum = 0;
   for (unsigned int i = 0; i < m_AxesO3.size(); ++i) {
     Sum += m_AxesO3[i].GetSum();
   }
