+++ src/global/misc/src/MResponseMatrixO6.cxx
@@ -769,11 +769,11 @@ float MResponseMatrixO6::GetMinimum() const
 ////////////////////////////////////////////////////////////////////////////////
 
 
-float MResponseMatrixO6::GetSum() const
+double MResponseMatrixO6::GetSum() const
 {
   // Return the sum of all bins:
 
-  float Sum = 0;
+  double Sum = 0;
   for (unsigned int i = 0; i < m_AxesO5.size(); ++i) {
     Sum += m_AxesO5[i].GetSum();
   }
